group platforms into chunks defined in data/chunks.ron
