+++ Cargo.toml
@@ -17,6 +17,8 @@ bevy_rapier2d = {version = "0.12.1", features = ["simd-nightly"]}
 bevy-inspector-egui = "0.8.2"
 bevy_loading = "0.2.0"
 rand = "0.8.5"
+ron = "0.6.4"
+serde = "1.0.127"
 
 [features]
 default = ["bevy/dynamic"] # to make the wasm github action work with the --no-default-feature flag
