+++ data/chunks.ron
@@ -1,0 +1,28 @@
+(
+	chunks: {
+		JumpPrelude: (
+			platforms: [
+				(
+					platform_kind: Platform1,
+					position: (0.0, -290.0),
+				),
+				(
+					platform_kind: Platform1,
+					position: (160.0, -290.0),
+				),
+				(
+					platform_kind: Platform1,
+					position: (320.0, -290.0),
+				),
+				(
+					platform_kind: Platform1,
+					position: (480.0, -290.0),
+				),
+				(
+					platform_kind: Platform1,
+					position: (640.0, -290.0),
+				)
+			]
+		)
+	}
+)
